Add files via upload
In this project, I created a visualization of a dataset that provided data for the price, product, and stores in which products were purchased from in preparation of a "zombie apocalypse". The visualization provides a sum of price per product for each store to determine where the most and least amount of money was spent for each store and a visualization for the best store to purchase each type of product based on the sum of the price.